Moderation mail stuffs

+++ Lilia/Modules/Utils/ModerationModuleUtils.cs
@@ -21,11 +21,55 @@ public static class ModerationModuleUtils
 		return (false, testRole);
 	}
 
-	public static async Task<IEnumerable<SocketUser>> GetMentionedUsersAsync(ShardedInteractionContext ctx, InteractiveService interactive)
+	public static async Task<IEnumerable<SocketUser>> GetMentionedUsersAsync(ShardedInteractionContext ctx, InteractiveService interactive,
+		bool deleteTriggerMessage = true)
 	{
 		var result = await interactive.NextMessageAsync(x => x.Channel.Id == ctx.Channel.Id && x.Author == ctx.User);
 		var memberList = result.IsSuccess ? result.Value?.MentionedUsers.Distinct() : new List<SocketUser>();
-		await result.Value!.DeleteAsync();
+
+		if (deleteTriggerMessage)
+		{
+			await result.Value!.DeleteAsync();
+		}
+
 		return memberList;
+	}
+
+	public static async Task<IEnumerable<SocketRole>> GetMentionedRolesAsync(ShardedInteractionContext ctx, InteractiveService interactive,
+		bool excludeEveryone = true, bool deleteTriggerMessage = true)
+	{
+		var result = await interactive.NextMessageAsync(x => x.Channel.Id == ctx.Channel.Id && x.Author == ctx.User);
+
+		var roleList = result.IsSuccess
+			? result.Value?.MentionedRoles.Distinct()
+			: new List<SocketRole>();
+
+		if (excludeEveryone)
+		{
+			roleList = roleList!.ToList().Where(role => role.IsEveryone);
+		}
+
+		if (deleteTriggerMessage)
+		{
+			await result.Value!.DeleteAsync();
+		}
+
+		return roleList;
+	}
+
+	public static async Task<IEnumerable<SocketGuildChannel>> GetMentionedChannelsAsync(ShardedInteractionContext ctx, InteractiveService interactive,
+		bool deleteTriggerMessage = true)
+	{
+		var result = await interactive.NextMessageAsync(x => x.Channel.Id == ctx.Channel.Id && x.Author == ctx.User);
+		var channelList = result.IsSuccess
+			? result.Value?.MentionedChannels.Distinct()
+			: new List<SocketGuildChannel>();
+
+		if (deleteTriggerMessage)
+		{
+			await result.Value!.DeleteAsync();
+		}
+
+		return channelList;
 	}
 }
